requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve   
langchain-groq
fastapi
uvicorn
sse_starlette
langchain
ipykernel
bs4
pypdf
langchain-text-splitters
sentence_transformers
langchain-huggingface
chromadb